backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.9
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extra-fields==3.0.2
django-filter==23.5
django-import-export==3.3.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docker==7.0.0
et-xmlfile==1.1.0
future==0.18.3
gunicorn==20.1.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
serializers==0.2.4
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
urllib3==2.1.0
webcolors==1.13
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
